[internationalization-implementation] Internationalization fix, remove unused files and implement SlideOverCard View
