--- PulseEcoSwiftUI/PulseEcoSwiftUI/PartialViews/BlurEffectView.swift
@@ -1,23 +1,0 @@
-//
-//  BlurEffectView.swift
-//  PulseEcoSwiftUI
-//
-//  Created by Maja Mitreska on 1/26/21.
-//  Copyright © 2021 Monika Dimitrova. All rights reserved.
-//
-
-import SwiftUI
-
-struct BlurEffectView: UIViewRepresentable {
-
-    /// The style of the Blut Effect View
-    var style: UIBlurEffect.Style = .systemMaterial
-
-    func makeUIView(context: Context) -> UIVisualEffectView {
-        return UIVisualEffectView(effect: UIBlurEffect(style: style))
-    }
-
-    func updateUIView(_ uiView: UIVisualEffectView, context: Context) {
-        uiView.effect = UIBlurEffect(style: style)
-    }
-}
